+++ PCA9955APWM.h
@@ -159,7 +159,7 @@ private:
   void init(void);
   uint8_t pwm_register_access(uint8_t port);
   uint8_t current_register_access(uint8_t port);
-  uint8_t simple_exp(uint8_t refIn);
+  float simple_exp(float refIn);
   static boolean _isMyDevice(uint8_t i2cAddr, TwoWire *i2cPort = &Wire);
 
   const uint8_t n_of_ports;
